refactor: renameed hooks
