+++ surfsense_web/app/dashboard/[search_space_id]/researcher/[[...chat_id]]/page.tsx
@@ -5,8 +5,8 @@ import { useParams, useRouter } from "next/navigation";
 import { useEffect, useMemo } from "react";
 import type { ResearchMode } from "@/components/chat";
 import ChatInterface from "@/components/chat/ChatInterface";
+import { useChatAPI, useChatState } from "@/hooks/use-chat";
 import type { Document } from "@/hooks/use-documents";
-import { useChatAPI, useChatState } from "@/hooks/useChat";
 
 export default function ResearcherPage() {
 	const { search_space_id, chat_id } = useParams();
