+++ surfsense_web/hooks/index.ts
@@ -1,3 +1,3 @@
 export * from "./use-document-by-chunk";
 export * from "./use-logs";
-export * from "./useSearchSourceConnectors";
+export * from "./use-search-source-connectors";
